+++ src/apply-filter.tsx
@@ -5,7 +5,7 @@
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
  * Created at     : 2023-07-06 14:52:33
- * Last modified  : 2023-07-06 15:47:50
+ * Last modified  : 2024-06-26 21:37:46
  */
 
 import { Action, ActionPanel, Grid } from "@raycast/api";