+++ src/components/SettingsActionPanelSection.tsx
@@ -1,3 +1,13 @@
+/**
+ * @file components/SettingsActionPanelSection.tsx
+ *
+ * @summary Action panel for opening command and extension preferences.
+ * @author Stephen Kaplan <skaplanofficial@gmail.com>
+ *
+ * Created at     : 2023-07-06 16:46:33
+ * Last modified  : 2024-06-26 21:37:46
+ */
+
 import { Action, ActionPanel, Icon, openCommandPreferences, openExtensionPreferences } from "@raycast/api";
 
 export default function SettingsActionPanelSection() {