+++ src/create-image.tsx
@@ -5,7 +5,7 @@
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
  * Created at     : 2023-07-06 14:53:50
- * Last modified  : 2023-07-06 16:48:08
+ * Last modified  : 2024-06-26 21:37:46
  */
 
 import { Color, Grid, LaunchProps, useNavigation } from "@raycast/api";