+++ src/operations/filterOperation.ts
@@ -5,7 +5,7 @@
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
  * Created at     : 2023-07-dd 00:32:49
- * Last modified  : 2023-07-dd 00:32:49
+ * Last modified  : 2024-06-26 21:37:46
  */
 
 import { Filter } from "../utilities/types";