+++ src/operations/rotateOperation.ts
@@ -5,7 +5,7 @@
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
  * Created at     : 2023-07-05 23:24:24
- * Last modified  : 2023-07-06 14:52:04
+ * Last modified  : 2024-06-26 21:37:46
  */
 
 import { execSync } from "child_process";