+++ src/utilities/clipboard.ts
@@ -5,7 +5,7 @@
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
  * Created at     : 2023-07-06 14:50:26
- * Last modified  : 2023-07-06 15:48:31
+ * Last modified  : 2024-06-26 21:37:46
  */
 
 import { runAppleScript } from "@raycast/utils";