+++ src/utilities/filters.ts
@@ -4,8 +4,8 @@
  * @summary Helper functions and resources for applying filters to images and PDFs using Core Image and ASObjC.
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
- * Created at     : 2023-07-dd 00:44:28
- * Last modified  : 2023-07-dd 00:44:28
+ * Created at     : 2023-07-06 00:44:28
+ * Last modified  : 2024-06-26 21:37:46
  */
 
 import { runAppleScript } from "@raycast/utils";