+++ src/utilities/preferences.ts
@@ -5,7 +5,7 @@
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
  * Created at     : 2023-07-06 15:34:53
- * Last modified  : 2023-07-06 15:41:56
+ * Last modified  : 2024-06-26 21:37:46
  */
 
 /**