+++ src/utilities/types.ts
@@ -5,7 +5,7 @@
  * @author Stephen Kaplan <skaplanofficial@gmail.com>
  *
  * Created at     : 2023-07-06 14:47:41
- * Last modified  : 2023-07-06 15:48:21
+ * Last modified  : 2024-06-26 21:37:46
  */
 
 /**